check if file exists before removing

+++ deploy/scripts/prepare_region.sh
@@ -264,10 +264,14 @@ if [ -n "${temp}" ]; then
     echo "#                                                                                       #"
     echo "#########################################################################################"
     echo ""
-    rm stdout.az
-    exit -1
+    if [ -f stdout.az ]
+        rm stdout.az
+    fi
+        exit -1
 else
-    rm stdout.az
+    if [ -f stdout.az ]
+        rm stdout.az
+    fi
 fi
 
 curdir=$(pwd)
